Fixed normalization logic
Normalization logic was not normalizing break segments, producing incorrect break lengths. Added BPM to the normalization breakdown. Also added a way to choose the best viable BPM in a given chart and use this for the normalization BPM. Added unit tests for normalization. (note that normalization may still be off by up to a measure, since we're just doing flat multiplication)